requirements.txt
Django>=6.0,<6.1
djangorestframework>=3.16,<3.17
djangorestframework-simplejwt>=5.5,<5.6
django-filter>=25.1,<25.3
drf-spectacular>=0.28,<0.29
psycopg[binary]>=3.2,<3.3
django-cors-headers>=4.7,<4.8
django-environ>=0.12,<0.13
Pillow>=11.2,<11.3
